Issue 013: Fix validation errors
Fixed these validation errors in the templates:
- Source tag under prolog in the wrong location: should come after
author.
- Glossary entry topic: only one type allowed for each glossAlt entry
(for the template the different types are included in a comment, with
instructions to choose one).
- No id values: added value 'replace-with-real-id' for all id
attributes.

For the glossary entry topic topic added prolog tags for consistency
with other templates (did not include metadata elements as per The IBM
Style Guide (page 259) - do not create index entries for glossary
terms).

For r-gpl-3.0.dita topic put source tag in correct location.